Tidied up some styles, and created a behavior input control
We need a rapid way to add behaviors. Right now it isn't hooked
up to anything, but the most important behavior, whereby you can type
and press enter and it will then add the behavior is set up.

Pressing enter in either the name, or the level clears the name, and
focuses back on the name field. This way a user can add loads of entries
without leaving the keyboard.